feat: ban confirmation
